- int -> long conversion for DB ids - double -> float conversion for values stored in netCDF - close inputstream in analytics, netcdfs in web app - netcdf store longs as doubles - rename atlasdao -> atlassolrdao - gene properties to lowercase (?!)
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@11568 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-dao/src/test/java/uk/ac/ebi/gxa/dao/procedures/SampleSetter.java
@@ -44,7 +44,7 @@ public class SampleSetter {
     Statement stmt = conn.createStatement();
 
     // create an sampleid - no oracle id generators here!
-    int sampleid = (int) System.currentTimeMillis();
+    long sampleid = (long) System.currentTimeMillis();
 
     stmt.executeQuery(
         "INSERT INTO A2_SAMPLE(sampleid, accession, species, channel) " +
