+++ atlas-model/src/main/java/uk/ac/ebi/microarray/atlas/model/Species.java
@@ -26,19 +26,19 @@ package uk.ac.ebi.microarray.atlas.model;
  * @author pashky
  */
 public class Species {
-    private int id;
+    private long id;
     private String name;
 
-    public Species(int id, String name) {
+    public Species(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
